Listitem pagination

+++ src/site/list/List.vue
@@ -3,37 +3,39 @@
         <SectionMain>
             <SectionTitleLineWithButton :icon="mdiAccount" title="Todo List" main />
             <CardBox :classType="'light'">
-                <table class="min-w-full border border-collapse border-gray-200 table-auto">
-                    <thead>
-                        <tr>
-                            <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left uppercase text-slate-900 bg-gray-50">ID</th>
-                            <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-900 uppercase bg-gray-50">Title</th>
-                            <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-900 uppercase bg-gray-50">Description</th>
-                            <th colspan="2" class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-900 uppercase bg-gray-50">Action</th>
-                        </tr>
-                    </thead>
-                    <tbody class="bg-white divide-y divide-gray-200">
-                        <tr v-for="(item, index) in listItems" :key="index">
-                            <td class="px-6 py-4 text-sm whitespace-no-wrap text-slate-800">
-                                {{ item.id }}
-                            </td>
-                            <td class="px-6 py-4 text-sm whitespace-no-wrap text-slate-800">
-                                {{ item.title }}
-                            </td>
-                            <td class="px-6 py-4 text-sm whitespace-no-wrap text-slate-800">
-                                {{ truncateDescription(item.description) }}
-                            </td>
-                            <td class="px-2 py-4 whitespace-no-wrap">
-                                <RouterLink :to="{ name: 'list.show', params: { list_id: item.id } }" class="text-sm text-indigo-600 hover:text-indigo-900">
-                                    Edit
-                                </RouterLink>
-                            </td>
-                            <td class="px-2 py-4 whitespace-no-wrap">
-                                <button @click="removeListItem(item.id)" class="text-sm text-red-600 hover:text-red-900">Delete</button>
-                            </td>
-                        </tr>
-                    </tbody>
-                </table>
+                <div class="genHeight h-450">
+                    <table class="min-w-full border border-collapse border-gray-200 table-auto">
+                        <thead>
+                            <tr>
+                                <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left uppercase text-slate-900 bg-gray-50">ID</th>
+                                <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-900 uppercase bg-gray-50">Title</th>
+                                <th class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-900 uppercase bg-gray-50">Description</th>
+                                <th colspan="2" class="px-6 py-3 text-xs font-medium leading-4 tracking-wider text-left text-gray-900 uppercase bg-gray-50">Action</th>
+                            </tr>
+                        </thead>
+                        <tbody class="bg-white divide-y divide-gray-200">
+                            <tr v-for="(item, index) in listItems" :key="index">
+                                <td class="px-6 py-4 text-sm whitespace-no-wrap text-slate-800">
+                                    {{ item.id }}
+                                </td>
+                                <td class="px-6 py-4 text-sm whitespace-no-wrap text-slate-800">
+                                    {{ item.title }}
+                                </td>
+                                <td class="px-6 py-4 text-sm whitespace-no-wrap text-slate-800">
+                                    {{ truncateDescription(item.description) }}
+                                </td>
+                                <td class="px-2 py-4 whitespace-no-wrap">
+                                    <RouterLink :to="{ name: 'list.show', params: { list_id: item.id } }" class="text-sm text-indigo-600 hover:text-indigo-900">
+                                        Edit
+                                    </RouterLink>
+                                </td>
+                                <td class="px-2 py-4 whitespace-no-wrap">
+                                    <button @click="removeListItem(item.id)" class="text-sm text-red-600 hover:text-red-900">Delete</button>
+                                </td>
+                            </tr>
+                        </tbody>
+                    </table>
+                </div>
                 <Paginate :next-page-url="nextPageUrl" type="load-more" @update:ModelValue="getListItems($event)" />
             </CardBox>
         </SectionMain>
@@ -42,9 +44,10 @@
 
 <script setup>
 import { onMounted, ref} from 'vue';
-import { mdiAccount, mdiAsterisk } from '@mdi/js';
+import { mdiAccount } from '@mdi/js';
 import apiRepository from "@/composables/apiRepository";
 import CommonFunction from "@/composables/CommonFunction";
+import Swal from 'sweetalert2';
 import CardBox from '@/components/CardBox.vue';
 import Paginate from "@/components/Paginate.vue";
 import SectionMain from '@/components/SectionMain.vue';
@@ -63,7 +66,6 @@ const nextPageUrl = ref(null);
 // Functions
 const getListItems = (url = null) => {
     getAllList(url).then((response) => {
-        console.log('response.data', response.data);
         const { data } = response;
         listItems.value.push(...data.items.data);
         nextPageUrl.value = data.items.next_page_url;
@@ -73,14 +75,32 @@ const getListItems = (url = null) => {
 };
 
 const removeListItem = (list_id) => {
-    processing.value = true;
-    
-    deleteList(list_id).then((response) => {
-        fireToaster(response.data.message, 'success');
-    }).catch((errors) => {
-        console.log(errors);  
-    }).finally(() => {
-        processing.value = false;
+
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            processing.value = true;
+            
+            deleteList(list_id).then((response) => {
+                if(response.status == 200) {
+                    fireToaster('List Item Deleted Successfully', 'success');
+                    listItems.value = listItems.value.filter((item) => item.id != list_id);
+                } else {
+                    fireToaster(response.data.message ? response.data.message : response.response.data.error[0], 'error');
+                }
+            }).catch((errors) => {
+                console.log(errors);  
+            }).finally(() => {
+                processing.value = false;
+            });
+        }
     });
 }
